Added possibility to give user acces to all commands
It could be a tedious job to add all the allowed permissions to the
config.yml file.

This line of code adds the ability to give a single user (or multiple)
acces to all command. Therefor is must be used with caution and only be
given to e.g. a server owner or a well trusted admin.

Usage in config.yml:
permissions:
  TelegramUsername:
    *: true